resume_agent_service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langchain>=0.1.0
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-groq
langchain-community>=0.0.10
httpx>=0.26.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
langchain-huggingface
sentence-transformers
langgraph-checkpoint
langgraph-checkpoint-sqlite
SQLAlchemy
pypdf
faiss-cpu
